moving dropdown from model
